requirements.txt
Django==4.1.3
django-admin-sortable2==2.1.4
django-tinymce==3.5.0
Pillow==9.3.0
python-dotenv==0.21.0
environs==9.5.0
